backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.12.0
elevenlabs==1.2.0
httpx==0.26.0
